Update Kubectl function to report errors
The Kubectl function in the test utils that we have been using does not
correctly report/fail if the command fails. This might be contributing
to some of the currently flakey tests in other PRs. This will at least
help in debugging the issues.

It is something that should be eventually addressed in the test util
package in another repo, but is easier to test it here like this, rather
than wait for it to be merged somewhere else and then update the
dependency.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>