Sync with midstream repository (#21)
* Organization rename to `stolostron` (#80)

- https://github.com/stolostron/backlog/issues/18774

Signed-off-by: Dale Haiducek <dhaiduce@redhat.com>

* Update golang.org/x/crypto (CVE-2021-43565) (#81)

Signed-off-by: Yu Cao <ycao@redhat.com>

* Fix lease creation (#91)

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Update Kubebuilder to 3.2.0 (#89)

This is in parity with `operator-sdk` 1.16.0:
https://github.com/operator-framework/operator-sdk/blob/v1.16.0/go.mod#L42

Signed-off-by: Dale Haiducek <dhaiduce@redhat.com>

* Update Go packages (#97)

Signed-off-by: Dale Haiducek <dhaiduce@redhat.com>

* Update `client-go` (#98)

Signed-off-by: Dale Haiducek <dhaiduce@redhat.com>

* Change addon lease name, remove old checks (#99)

The new policy addon controller splits the previous "policy-controller"
addon into separate "config-policy-controller" and "governance-policy-
framework" addons. Each of those addons can manage their own lease, and
this controller no longer needs to check on the health of the config
policy controller.

Refs:
 - https://github.com/stolostron/backlog/issues/19521
 - https://github.com/stolostron/backlog/issues/20013
 - https://github.com/stolostron/backlog/issues/11508

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Fix label selector in addon lease health check (#100)

The pod is deployed with label "governance-policy-framework" in the
addon chart. This updates that label in this repo's deployment YAML, as
well as in the health check for the addon lease.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Configure logging with utility package (#101)

This will make the logs look more consistent, and allow for better run-
time configuration of both logs this controller creates, and the logs
that libraries might create.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Remove namespace creation (#102)

* Remove namespace creation

This will now be handled in the addon deployment.

Refs:
- https://github.com/stolostron/governance-policy-addon-controller/pull/24

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Remove namespace creation tests

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Remove namespace permissions (#103)

These permissions are no longer required since the cluster namespace is
not managed by this controller.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Sync the upstream repo

Changes made to midstream are mostly build and logging related.  This
updates changes from midstream to our upstream repository.

Refs:
 - https://github.com/stolostron/backlog/issues/21103

Signed-off-by: Gus Parvin <gparvin@redhat.com>

Co-authored-by: Dale Haiducek <dhaiduce@redhat.com>
Co-authored-by: Yu Cao <ycao@redhat.com>
Co-authored-by: Justin Kulikauskas <44813129+JustinKuli@users.noreply.github.com>