Fix deploy (#107)
* Restore klog init

Otherwise, the controller crashes if provided with the `--v=0` flag,
which we have in the operator.yaml.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add deployment YAML verification

Since the tests are now run through a local "instrumented" binary for
test coverage, the operator.yaml is no longer used in our tests. This
ensures that the operator.yaml can successfully deploy.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>